apply PR feedback (#2822)
