+++ Dockerfile.puppeteer
@@ -1,8 +1,5 @@
 FROM node:8-slim
 
-# See https://crbug.com/795759
-RUN apt-get update && apt-get install -yq libgconf-2-4
-
 # Install latest chrome dev package and fonts to support major charsets (Chinese, Japanese, Arabic, Hebrew, Thai and a few others)
 # Note: this installs the necessary libs to make the bundled version of Chromium that Puppeteer
 # installs, work.
