+++ app/assets/javascripts/test/puppeteer/dataset_rendering_helpers.js
