+++ package.json
@@ -78,7 +78,7 @@
     "test-prepare": "tools/test.sh prepare",
     "test-prepare-watch": "tools/test.sh prepare -w",
     "test-e2e": "tools/postgres/prepareTestDB.sh && tools/test.sh test-e2e --timeout=60s --verbose",
-    "test-screenshot": "docker-compose run screenshot-tests bash -c 'node_modules/.bin/ava app/assets/javascripts/test/tracing-view-screenshot-tests/*.screenshot.js'",
+    "test-screenshot": "ava app/assets/javascripts/test/puppeteer/*.screenshot.js --timeout=5m",
     "test-help": "echo For development it is recommended to run yarn test-prepare-watch in one terminal and yarn test-watch in another. This is the fastest way to perform incremental testing. If you are only interested in running test once, use yarn test.",
     "remove-snapshots": "rm -r app/assets/javascripts/test/snapshots/public",
     "create-snapshots": "rm -r app/assets/javascripts/test/snapshots/public && docker-compose up e2e-tests",
